requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
Automat==0.8.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
Click==7.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==36.0.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
distro==1.4.0
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-import-export==2.6.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
doq==0.7.0
drf-yasg==1.20.0
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==3.3
importlib-metadata==1.5.0
importlib-resources==5.4.0
incremental==16.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonpatch==1.22
jsonpointer==2.0
jsonschema==4.2.1
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
MarkupPy==1.14
MarkupSafe==2.0.1
more-itertools==4.2.0
msgpack==0.6.2
Naked==0.1.31
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
parso==0.8.2
pbr==5.8.0
pexpect==4.6.0
protobuf==3.6.1
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
PyHamcrest==1.9.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.4.0
pynvim==0.4.1
pyOpenSSL==19.0.0
pyparsing==3.0.6
pyRFC3339==1.1
pyrsistent==0.18.0
pyserial==3.4
python-nmap==0.7.1
pytz==2021.3
pywin32-ctypes==0.2.0
pyxdg==0.26
PyYAML==6.0
requests==2.26.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scapy==2.4.5
SecretStorage==2.3.1
service-identity==18.1.0
shellescape==3.8.1
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.2
ssh-import-id==5.10
swagger-spec-validator==2.7.4
tablib==3.1.0
testresources==2.0.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.4.7
wadllib==1.3.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
zope.interface==5.4.0